keep builder slimmer and use expressions with appropriate constants
